Refactor mcp_router.py to package-based imports
Replaces the monolithic mcp_router.py implementation with a backwards-compatibility shim that re-exports all symbols from the new scripts.mcp_router package. This enables modularization and easier maintenance while preserving import and CLI compatibility.
