+++ scripts/mcp_router/__main__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+"""
+Allow running as: python -m scripts.mcp_router "query"
+"""
+from .cli import main
+
+if __name__ == "__main__":
+    raise SystemExit(main())
